Add files via upload
Implementing an embedded system inside a bank which monitors the clients queue in front of the tellers. Monitoring was divided into calculating number of people waiting in the queue, and displaying the expected waiting time for each client in the queue, and the maximum number of clients in the queue. The system was simulated using ModelSim, and coded using Verilog.